comments and renaming files for consistency with other files
