+++ backend/routes/email.router.js
@@ -1,7 +1,18 @@
 import express from 'express';
-const router = express.Router();
 import { sendEmail } from '../controllers/email.controller.js';
 
+/**
+ * Email routes configuration
+ * Handles HTTP endpoints related to email communication
+ * Provides functionality for sending emails from the platform
+ */
+const router = express.Router();
+
+/**
+ * Send email to recipients
+ * @route POST /api/email
+ * @access Public - Accessible by any authenticated user
+ */
 router.post('/', sendEmail);
 
 export default router;
