+++ frontend/package.json
@@ -30,6 +30,7 @@
     "react-chartjs-2": "^5.3.0",
     "react-dom": "^18.3.1",
     "react-icons": "^5.5.0",
+    "react-paginate": "^8.3.0",
     "react-router-dom": "^7.1.3"
   },
   "devDependencies": {
